ai-agent/requirements.txt
# WebSocket
websockets>=12.0

# VAD
webrtcvad>=2.0.10

# -----------------------------------------------------------------------------
# ASR (Automatic Speech Recognition)
# -----------------------------------------------------------------------------
# faster-whisper: Recomendado - muito mais rápido que whisper original
faster-whisper>=1.0.0

# whisper original (fallback)
# openai-whisper>=20231117

# -----------------------------------------------------------------------------
# LLM (Large Language Model)
# -----------------------------------------------------------------------------
anthropic>=0.18.0
openai>=1.0.0

# -----------------------------------------------------------------------------
# TTS (Text-to-Speech)
# -----------------------------------------------------------------------------
# Kokoro: TTS neural local de alta qualidade
kokoro>=0.3.0
soundfile>=0.12.0

# gTTS (fallback - gratuito mas requer internet)
gtts>=2.5.0

# -----------------------------------------------------------------------------
# Utils
# -----------------------------------------------------------------------------
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Metrics
prometheus-client>=0.19.0
